Mechanic mines (#26)
Add options for SF2e Mechanic's mines.
It triggers when the "Deploy Mine" or the "Double Deployment" actions are posted to chat

Notes:
An actor has been added to the Actor's compendium. Be careful before merging :)

The actor has base features for damage computation: a Deploy Mine action, a few rule elements to select the damage type, a "Critical Explosion" passive action that is added to the actor if the Mechanic has the Critical Explosion class feat.

Mine token art by myself.